requirements.txt

cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
mpmath==1.2.1
numpy==1.20.1
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
Pint==0.16.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.1
six==1.15.0
sympy==1.7.1
modsimpy==1.1.3